+++ inst/pipelines/use_targets.R
@@ -1,0 +1,37 @@
+# Follow the comments below to fill in this target script.
+# Then follow the manual to check and run the pipeline:
+#   https://books.ropensci.org/targets/walkthrough.html#inspect-the-pipeline # nolint
+
+# Load packages required to define the pipeline:
+library(targets)
+# library(tarchetypes) # Load other packages as needed. # nolint
+
+# Set target options:
+tar_option_set(
+  packages = c("tibble"), # packages that your targets need to run
+  format = "rds" # default storage format
+  # Set other options as needed.
+)
+
+# Configure the backend of tar_make_clustermq() (recommended):
+CLUSTERMQ
+
+# Configure the backend of tar_make_future() (optional):
+FUTURE
+
+# Load the R scripts with your custom functions:
+for (file in list.files("R", full.names = TRUE)) source(file)
+# source("other_functions.R") # Source other scripts as needed. # nolint
+
+# Replace the target list below with your own:
+list(
+  tar_target(
+    name = data,
+    command = tibble(x = rnorm(100), y = rnorm(100))
+#   format = "feather" # efficient storage of large data frames # nolint
+  ),
+  tar_target(
+    name = model,
+    command = coefficients(lm(y ~ x, data = data))
+  )
+)